option to build lib without tiff
