+++ src/tiff.h
@@ -17,4 +17,16 @@ class FreeDeleter {
 
 using TiffPtr = std::unique_ptr<TIFF, CloseDeleter>;
 
+struct GeoTIFFInfo {
+  double XGeoRef, YGeoRef;
+  double XCellRes, YCellRes;
+  double projection[16];
+  int nodata;
+  bool set;
+};
+
+int geotiff_read(TIFF* tiff, GeoTIFFInfo* info);
+
+int geotiff_write(TIFF* tiff, GeoTIFFInfo* info);
+
 }  // namespace multiblend::io::tiff
